Chore: #37-Onboarding 대대적 수정
