+++ Spha/Spha/SphaApp.swift
@@ -16,8 +16,7 @@ struct SphaApp: App {
     var body: some Scene {
         WindowGroup {
             NavigationStack(path: $router.path){
-                 MainView()
-//                DailyStatisticsView()
+                 OnboardingStartView()
                     .navigationDestination(for: SphaView.self){ sphaView in
                         router.view(for: sphaView)
                     }
